Rename CLI `dango auth` to `dango oauth` for Phase 2 auth
The `dango auth` command group previously handled OAuth credential
management (Google, Facebook, Shopify). Phase 2 needs `dango auth`
for user authentication (login, logout, roles). This renames the
OAuth commands to `dango oauth` and creates a placeholder `auth`
group for TASK-013.

- Rename cli/commands/auth.py → oauth.py (git mv + update internals)
- Create new placeholder cli/commands/auth.py for Phase 2
- Register both `auth` and `oauth` groups in cli/main.py
- Update ~35 "dango auth" string references across 8 source files
- Update tests: add oauth_subcommands test, auth placeholder test
- Update docs: CLAUDE.md files, ARCHITECTURE.md, file-exemptions.yml
