+++ dango/cli/CLAUDE.md
@@ -15,7 +15,8 @@ Click-based command-line interface for all Dango operations — project init, so
 | `commands/project.py` (292 lines) | `init`, `rename`, `info` | `init()`, `rename()`, `info()` |
 | `commands/source.py` (521 lines) | `source` group (`add`, `list`, `remove`) + `sync` | `source`, `sync()` |
 | `commands/platform.py` (955 lines) | `start`, `stop`, `status` | `start()`, `stop()`, `status()` |
-| `commands/auth.py` (707 lines) | `auth` group (10 subcommands) | `auth`, `auth_setup()`, `auth_status()`, `auth_check()`, etc. |
+| `commands/auth.py` (~20 lines) | `auth` group placeholder (Phase 2 user management) | `auth` |
+| `commands/oauth.py` (815 lines) | `oauth` group (10 subcommands) | `oauth`, `auth_setup()`, `auth_status()`, `auth_check()`, etc. |
 | `commands/transform.py` (326 lines) | `run`, `docs`, `generate` | `run()`, `docs()`, `generate()` |
 | `commands/data.py` (360 lines) | `db` group (`status`, `clean`) + `validate` | `db`, `validate()` |
 | `commands/config_cmd.py` (179 lines) | `config` group (`validate`, `show`) | `config` |
@@ -58,7 +59,8 @@ dango (top-level group)
 │   ├── validate, show
 ├── db (group)                  ← commands/data.py
 │   ├── status, clean
-├── auth (group)                ← commands/auth.py
+├── auth (group)                ← commands/auth.py (Phase 2 placeholder)
+├── oauth (group)               ← commands/oauth.py
 │   ├── status, setup, check, list, remove, refresh,
 │   │   facebook_ads, google_sheets, google_analytics, google_ads
 ├── model (group)               ← commands/model.py
