+++ dango/cli/main.py
@@ -16,6 +16,7 @@ from dango.cli.commands.data import db, validate
 from dango.cli.commands.metabase_cmd import metabase
 from dango.cli.commands.migrate import migrate
 from dango.cli.commands.model import model
+from dango.cli.commands.oauth import oauth
 from dango.cli.commands.platform import start, status, stop
 from dango.cli.commands.project import info, init, rename
 from dango.cli.commands.source import source, sync
@@ -73,6 +74,7 @@ cli.add_command(source)
 cli.add_command(config)
 cli.add_command(db)
 cli.add_command(auth)
+cli.add_command(oauth)
 cli.add_command(model)
 cli.add_command(dashboard)
 cli.add_command(metabase)
