+++ dango/cli/oauth.py
@@ -419,9 +419,9 @@ def check_token_expiry(project_root: Path, source_type: str) -> str | None:
                         days_remaining = (expiry_date - datetime.now()).days
 
                         if days_remaining < 0:
-                            return f"[red]Facebook token expired {abs(days_remaining)} days ago[/red]\nRun: dango auth facebook_ads"
+                            return f"[red]Facebook token expired {abs(days_remaining)} days ago[/red]\nRun: dango oauth facebook_ads"
                         elif days_remaining < 7:
-                            return f"[yellow]Facebook token expires in {days_remaining} days[/yellow]\nConsider running: dango auth facebook_ads"
+                            return f"[yellow]Facebook token expires in {days_remaining} days[/yellow]\nConsider running: dango oauth facebook_ads"
                 except Exception:
                     pass
