+++ docs/file-exemptions.yml
@@ -30,10 +30,10 @@ exemptions:
     reason: "Platform lifecycle commands (start/stop/status) + port helpers. Extracted from cli/main.py by TASK-005. Further splitting would scatter tightly coupled startup/shutdown logic."
     review_by: "v1.1"
 
-  - file: dango/cli/commands/auth.py
+  - file: dango/cli/commands/oauth.py
     lines: 815
     category: exempt
-    reason: "OAuth authentication commands (10 subcommands). Extracted from cli/main.py by TASK-005. One command per provider — file grows linearly with providers."
+    reason: "OAuth authentication commands (10 subcommands). Renamed from auth.py by TASK-093. One command per provider — file grows linearly with providers."
     review_by: "v1.1"
 
   - file: dango/cli/commands/source.py
